Bleh!

+++ kernel/stdlib/stddef.h
@@ -2,6 +2,6 @@
 
 #define NULL (void*)0
 typedef unsigned long size_t;
-typedef size_t ptrdiff_t;
+typedef signed long ptrdiff_t;
 
 #define offsetof(st, m) ((size_t)&(((st *)0)->m))
